refactor: organise code
