+++ src/stats.ts
@@ -1,6 +1,6 @@
-import { VarnaType, TokenType } from "@vipran/aksharas";
+import Aksharas, { VarnaType, TokenType } from "@vipran/aksharas";
 
-export interface Stats {
+interface Stats {
   total: number;
   totalAksharas: number;
   totalChars: number;
@@ -13,13 +13,30 @@ export interface Stats {
   totalVyanjanas: number;
 }
 
+const statEles = [
+  "counter-total",
+  "counter-total-aksharas",
+  "counter-total-symbols",
+  "counter-total-others",
+  "counter-total-varnas",
+  "counter-total-svaras",
+  "counter-total-vyanjanas",
+  "counter-total-spaces",
+  "counter-total-chars",
+  "counter-total-invalids",
+].map((className) =>
+  document.querySelector<HTMLSpanElement>(
+    `#app .stats .${className} .counter-content`
+  )
+);
+
 const filterVarnas = (varnas: any[], varnaType: VarnaType) =>
   varnas.filter((varna) => varna.type === varnaType);
 
 const filterSpaces = (tokens: any[]) =>
   tokens.filter((token) => token.value === " ");
 
-export const getStats = (results: any): Stats => {
+const getStats = (results: any): Stats => {
   const validTokens = results.all.filter(
     (token: any) => token.type !== TokenType.Invalid
   );
@@ -36,4 +53,31 @@ export const getStats = (results: any): Stats => {
     totalVarnas: results.varnas.length,
     totalVyanjanas: filterVarnas(results.varnas, VarnaType.Vyanjana).length,
   };
+};
+
+const $renderStats = (stats: Stats) => {
+  const $stats = [
+    stats.total,
+    stats.totalAksharas,
+    stats.totalSymbols,
+    stats.totalOtherChars,
+    stats.totalVarnas,
+    stats.totalSvaras,
+    stats.totalVyanjanas,
+    stats.totalSpaces,
+    stats.totalChars,
+    stats.totalInvalidChars,
+  ];
+
+  statEles.forEach(
+    (statEle, i) => statEle && (statEle.innerHTML = $stats[i].toString())
+  );
+};
+
+export const renderStats = (input: string) => {
+  const results = Aksharas.analyse(input);
+
+  const stats = getStats(results);
+
+  $renderStats(stats);
 };
